Added text output files for the lines printed to stdout for each of 3 cases of command line args: -numclients=1000 -timeout=90 -delay=150 -numclients=1000 -timeout=90 -delay=200 -numclients=1000 -timeout=90 -delay=250
It's noticeable, and odd, that the missing goroutines to be completed are "close to multiples" of the delay factor.
So for 150:
150
300
450
600
750
900

This smells bad. Something (that I can't yet explain) is going wrong here. Implementing the wg.wait at line 76 would simply cause the program to stall
From previous tests I've seen that the effect of the time delay inside the goroutines is that ALL the goroutines get launched v. quickly -  and groups (on each multiple) get released in sequence.