Moved time.sleep for each client outside the goroutine into the surrounding loop
This has the effect of spacing out the messages.
The code also adds the debug flag for "basichost". This makes the debug messages for p2p/host/basic/basic_host.go appear
This still produces problems with the numbers of completed connections being less than the number
requested.
I altered the go-libp2p to use a local copy so I could change the debug message on line 735 of basic_host.go to "...start dialing"..
This allowed anb easy count of occurrences vs the numbers of "...finished dialing".

The number of completed connections disappears if both the select block from line 747 to line 751 is removed and the ping.
I then inserted a deliberate call h.ids.IdentifyConn(c) - so the portion in basic_host.go becomes.

`........
// TODO: Consider removing this? On one hand, it's nice because we can
	// assume that things like the agent version are usually set when this
	// returns. On the other hand, we don't _really_ need to wait for this.
	//
	// This is mostly here to preserve existing behavior.
	h.ids.IdentifyConn(c)
	select {
	case <-h.ids.IdentifyWait(c):
	case <-ctx.Done():
		return ctx.Err()
	}

...........`

The missing number of connections was then exact (see DialingCompare4.txt) - of the 1000 clients requested
Start Dialing - 1000
Finished Dialing - 853
NumWaitGroups - 147

The comments in github.com/gpestana/go-libp2p/blob/934606d0f292/p2p/protocol/identify/id.go indicate
that the IdentifyWait returns if and ONLY if the ProtocolIdentify (handshake3) finishes on the given connection.

So a problem with issuing/received ProtocolIdentify (handshake3) would appear to exist when spawning large numbers of clients rapidly.