Using Marco's suggested code as the basis..
Added extra logging for a local copy of the golibp2p library
Experimented...
Findings thus far:

Closing the client (at the end of newClient) gets rid of the problem
Setting the IP listen address deliberately to 127.0.0.1 deliberately improves the situation slightly - it can manage 100 connections but not many more

There large numbers of ErrReset issued from transport/quic/stream.go in the Read and Write Functions - which are seen as "stream reset" on the client end

At a higher level the error appears within the client call to SelectProtoOrFail. The actual error occurs in quic-go/receive_stream.go as a result of cancelRead.

Since the server is still alive and clients are also alive (even if quiet) I'm not sure why these errors are happening?